Cambio al Entity

+++ Entity/Cancha.cs
@@ -30,7 +30,7 @@ namespace Entity
 
         [Required(ErrorMessage = "Porfavor, ingrese un precio por hora: ")]
         [DisplayName("Precio x Hora")]
-        public float MPrecioHora { get; set; }
+        public decimal MPrecioHora { get; set; }
        
 }
 }
